refactor: extract Notes component and add dependencies
